pyproject.toml
[project]
name = "support-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "chromadb>=1.0.20",
  "google-generativeai>=0.8.5",
  "pypdf>=6.0.0",
  "sentence-transformers>=5.1.0",
  "webvtt-py>=0.5.1",
]

[project.optional-dependencies]
dev = ["black", "ruff"]

[project.scripts]
rag-cli = "rag_cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
  "E", # pycodestyle errors
  "W", # pycodestyle warnings
  "F", # pyflakes
  "I", # isort
  "C", # flake8-comprehensions
  "B", # flake8-bugbear
]
ignore = ["E501"] # let black handle line length
